refactor: refactor header
