+++ async-race/src/modules/components/textInput.ts
@@ -1,6 +1,6 @@
 export const createTextInput = (className: string): HTMLInputElement => {
   const input = document.createElement('input');
-  input.classList.add('input');
+  input.classList.add('input-text');
   input.classList.add(className);
 
   return input;
