+++ async-race/src/modules/features/index.ts
@@ -1,2 +1,6 @@
+import { createRaceHeaderBtns } from './raceHeaderBtns';
+
 export { createInputBlock } from './inputsBlock';
 export { createNavigation } from './navigation';
+export { createRaceHeaderBtns } from './raceHeaderBtns';
+export { createHeader } from './header';
