+++ async-race/src/modules/pages/garage.ts
@@ -1,22 +1,11 @@
-import { createInputBlock, createNavigation } from '../features';
-import { createRaceHeaderBtn } from '../features/raceHeaderBtn';
+import { createHeader } from '../features';
 
 export const draw = (): void => {
   const root = document.querySelector('.root') as HTMLElement;
 
   const fragment = document.createDocumentFragment();
 
-  const header = document.createElement('header');
-  header.classList.add('header');
-
-  const nav = createNavigation();
-  header.appendChild(nav);
-
-  const inputBlock = createInputBlock();
-  header.appendChild(inputBlock);
-
-  const headerRaceBtns = createRaceHeaderBtn();
-  header.appendChild(headerRaceBtns);
+  const header = createHeader();
 
   fragment.appendChild(header);
   root.appendChild(fragment);
